--- src/scripts/script.js
@@ -1,8 +1,0 @@
-import './card.js';
-import './cardlist.js';
-import './popUp.js';
-import './api.js';
-import './formvalidator.js';
-import './userinfo.js';
-
-
